fix/added navbar in all the webpages
-Created navbar component for all the pages
-Removed navbar section from the homepages' file
