+++ src/components/Navbar/Navbar.css
@@ -1,0 +1,5 @@
+.navbar {
+    height: 10vh;
+    min-height: 60px;
+  }
+
