+++ src/components/PrivateHomePage/PrivateHomePage.css
@@ -1,7 +1,8 @@
 .background {
     background: url("https://cdn.pixabay.com/photo/2021/02/28/09/38/food-6056720_1280.jpg") no-repeat center center;
     background-size: cover;
-    height: 100vh;
+    width: 100%;
+    height: 100%;
     display: flex;
     align-items: center;
     justify-content: center;
@@ -25,3 +26,9 @@
     background-color: #ff964f;
   }
   
+  .private-home{
+    height: 90vh;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
