+++ src/components/PublicHomePage/PublicHomePage.css
@@ -36,3 +36,14 @@
   .login-btn:hover {
     background-color: #ff964f;
   }
+
+  .public-home {
+    height: 90vh;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+  .background {
+    width: 100%;
+    height: 100%;
+  }
